refactor: 이동하는 함수 분리
